Extracting base client class - fixing tests
