+++ pykusto/_src/client.py
@@ -8,8 +8,8 @@ from azure.kusto.data.exceptions import KustoServiceError
 from azure.kusto.data.helpers import dataframe_from_result_table
 from azure.kusto.data.response import KustoResponseDataSet
 
-from pykusto import KQL
-from pykusto._src.client_base import KustoResponseBase, PyKustoClientBase, RetryConfig, NO_RETRIES
+from .client_base import KustoResponseBase, PyKustoClientBase, RetryConfig, NO_RETRIES
+from .kql_converters import KQL
 
 
 class KustoResponse(KustoResponseBase):
@@ -34,6 +34,7 @@ class PyKustoClient(PyKustoClientBase):
     __client: KustoClient
     __auth_method: Callable[[str], KustoConnectionStringBuilder]
 
+    # Static members
     __global_client_cache: Dict[str, KustoClient] = {}
     __global_cache_lock: Lock = Lock()
 
@@ -51,36 +52,39 @@ class PyKustoClient(PyKustoClientBase):
         :param auth_method: A method that returns a KustoConnectionStringBuilder for authentication. The default is 'KustoConnectionStringBuilder.with_az_cli_authentication'.
         A popular alternative is 'KustoConnectionStringBuilder.with_aad_device_authentication'
         """
+        self.__auth_method = auth_method
+        client_resolved = False
         if isinstance(client_or_cluster, KustoClient):
             self.__client = client_or_cluster
+            client_resolved = True
             # noinspection PyProtectedMember
             cluster_name = urlparse(client_or_cluster._query_endpoint).netloc
             assert not use_global_cache, "Global cache not supported when providing your own client instance"
         else:
             cluster_name = client_or_cluster
-            self.__client = (self._cached_get_client_for_cluster if use_global_cache else self._get_client_for_cluster)()
-        self.__auth_method = auth_method
         super().__init__(cluster_name, fetch_by_default, retry_config.retry_on(KustoServiceError))
+        if not client_resolved:
+            self.__client = (self._cached_get_client_for_cluster if use_global_cache else self._get_client_for_cluster)()
 
     def __repr__(self) -> str:
-        return f"PyKustoClient('{self.__cluster_name}')"
+        return f"PyKustoClient('{self._cluster_name}')"
 
     def _internal_execute(self, database: str, query: KQL, properties: ClientRequestProperties = None, retry_config: RetryConfig = None) -> KustoResponse:
-        resolved_retry_config = self.__retry_config if retry_config is None else retry_config
+        resolved_retry_config = self._retry_config if retry_config is None else retry_config
         return KustoResponse(resolved_retry_config.retry(lambda: self.__client.execute(database, query, properties)))
 
     def _get_client_for_cluster(self) -> KustoClient:
-        return KustoClient(self.__auth_method(self.__cluster_name))
+        return KustoClient(self.__auth_method(self._cluster_name))
 
     def _cached_get_client_for_cluster(self) -> KustoClient:
         """
         Provided for convenience during development, not recommended for general use.
         """
         with PyKustoClient.__global_cache_lock:
-            client = PyKustoClient.__global_client_cache.get(self.__cluster_name)
+            client = PyKustoClient.__global_client_cache.get(self._cluster_name)
             if client is None:
                 client = self._get_client_for_cluster()
-                PyKustoClient.__global_client_cache[self.__cluster_name] = client
+                PyKustoClient.__global_client_cache[self._cluster_name] = client
                 assert len(PyKustoClient.__global_client_cache) <= 1024, "Global client cache cannot exceed size of 1024"
 
         return client
