+++ pykusto/_src/query.py
@@ -5,7 +5,7 @@ from os import linesep
 from types import FunctionType
 from typing import Tuple, List, Union, Optional
 
-from .client import Table, KustoResponse, RetryConfig
+from .client_base import Table, KustoResponseBase, RetryConfig
 from .enums import Order, Nulls, JoinKind, Distribution, BagExpansion
 from .expressions import BooleanType, ExpressionType, AggregationExpression, _AssignmentBase, _AssignmentFromAggregationToColumn, _AssignmentToSingleColumn, \
     _AnyTypeColumn, BaseExpression, _AssignmentFromColumnToColumn, AnyExpression, _to_kql, _expression_to_type, BaseColumn, NumberType, OrderedType
@@ -280,7 +280,7 @@ class Query:
             kql = KQL(kql.replace(" |", linesep + "|"))
         return kql
 
-    def execute(self, table: Table = None, retry_config: RetryConfig = None) -> KustoResponse:
+    def execute(self, table: Table = None, retry_config: RetryConfig = None) -> KustoResponseBase:
         if self.get_table() is None:
             if table is None:
                 raise RuntimeError("No table supplied")
